Address problems with our `pre-commit` config
First, the reference to `gitlab.com` for `flake8` was incorrect.  It has
been corrected to `github.com`.

If one starts using a different python interpreter (for example, if you
upgrade your linux distribution and that upgrade defaults to a new
version) then your previous pre-commit setup will need to be re-
installed.

Because of the `gitlab.com` reference that was not possible.

The second change is to bump the versions of all the pre-commit programs
used, `black`, `flake8`, and `isort`, to their latest versions.  It is
unfortunate but the existing versions we are using don't work with newer
versions of the Python 3, in particular, Python 3.11, but they do work
with Python 3.9 and later.