Drop `podman kube play` in favor of `podman run`
The `load_keycloak.sh` script is extended to do the following:

 * Wait for the Keycloak health endpoint to come up
 * Run the custom configuration

The `run-pbench-in-a-can` script is updated to run the container
services independently with the `podman run` command.  The Pbench Server
container is moved out of the the pod and only dependant services now
run inside the pod.  The Pbench Server container communicates with the
pod's services using host networking.