Drop need for external `BASE_IMAGE` definition
Prior to this change, if one tried to use `jenkins/runlocal` without
setting `BASE_IMAGE`, the default of `UBI:9` would result in an error
if the system on which this script was invoked did not have a RHEL
entitlement.

This change simplifies things so that the RPM built is truly local to
the environment, and using the installed environment, the `BASE_IMAGE`
is to work with that RPM.

For example, if running on a Fedora 37 box, a Fedora 37 RPM would be
built and installed using the `fedora:37` base image.