Only perform "hard" cleanup on INT, QUIT, and TERM
If a functional test is successful, and `--cleanup` has been requested,
we will have already removed the POD and container.  There is not need
to also run the "hard" cleanup script.

Further, we emit a note when `--cleanup` is not requested after the
tests have finished to alert the user that the POD and container are
still running.

We also emit the print statements for the clean up being performed.