Pass long `--cleanup`; support only the 9s
All arguments to `jenkins/runlocal` are passed along to
`jenkins/run-server-unit-tests` as a simple way to support the
`--cleanup` option on `jenkins/runlocal`

Further, we only support defaulting to a `ubi9:latest` or
`centos:stream9` image in those respective environments.

Lastely, we add support for `BASE_IMAGE` to be defined ahead of time.