Alembic schema changes for Operation table
NOTE: After working with Pete get get alembic to successfully generate a revision
file for my changes, we realized what a pain it would be to migrate (and
test) an actual live database. We need to have a functional test framework to
stand up an "old" functional DB, upgrade it to the latest revision, and then
validate the correctness.

Pending that effort, it makes little sense at this point to upgrade the
staging server database. Instead, we'll re-deploy.