Build and run functional tests all locally
The `jenkins/runlocal` script now builds a Pbench Server RPM from the
checked-out sources in the current working tree, builds the Pbench
Server CI container, and runs the functional tests against the CI
container image.

As a result, `jenkins/runlocal` no longer pushes images to a registry.

It was necessary to support alternate base images for server container
to allow for fully local builds.  Prior to this change, if one tried to
use `jenkins/runlocal` without setting `BASE_IMAGE`, the default of
`UBI:9` would result in an error if the system on which this script was
invoked did not have a RHEL entitlement.

This change simplifies things so that the RPM built is truly local to
the environment, and using the installed environment, derives the
`BASE_IMAGE` is to work with the built RPM.

For example, if running on a Fedora 37 box, a Fedora 37 RPM would be
built and installed using the `fedora:37` base image.

NOTE, we only support defaulting to a `ubi9:latest` or `centos:stream9`
image in those respective environments.

We now only perform "hard" cleanup on INT, QUIT, and TERM signals.  If a
functional test is successful, and `--cleanup` has been requested, we
will have already removed the POD and container.  There is no need to
also run the "hard" cleanup script.

Further, we emit a note when `--cleanup` is not requested after the
functional tests have finished to alert the user that the POD and
container are still running.  We also emit the print statements for the
clean up being performed.

All arguments to `jenkins/runlocal` are passed along to
`jenkins/run-server-unit-tests` as a simple way to support the
`--cleanup` option on `jenkins/runlocal`.

We also:

 * Fail running server functional tests if pod creation fails
 * Remove the need for the `jenkins/runner`
 * Remove the unnecessary use of `--format=docker` in the `buildah
   from` command
 * Drop specific linting references, add `isort` in `README.md`
 * Use local file to keep CI container image name
   This allows us to entirely remove all references to a specific
   container registry from the code so that developers can create a
   local file in which to store the registry want to use in parallel
   with what the CI jobs already do with Jenkins credentials.
 * Correct a comment in `jenkins/Pipeline.gy`
 * Correct `GITTOP` definition location
 * Use `/bin/bash -e` consistently, adding it to `run-pbench-in-a-can`