Add expirations on containers pushed to remote registries (#3230)
This PR modifies the places where the CI pushes containers to the remote registry to also set an automatic expiration date on the pushed containers. In this way, containers pushed by PR builds will eventually be removed without requiring manual clean-up of the remote registry. This change only affects Agent and Server containers; it doesn't affect the persistent containers such as the CI's own container and the containers used to build RPMs; it also doesn't affect branch builds which are "published" by the CI, and it doesn't affect containers which are "released". It only affects containers which are built by the CI for functional testing.

Details:
- Add a script which is basically just a wrapper around a curl command which hits a Quay RESTful interface to set the expiration date for the container that we just pushed. This implicitly assumes that the remote registry is running an implementation of Quay (which each of our remote registries currently does).
- Removed the "transport" specifier from the default value for IMAGE_REPO in the Agent container makefile. This is the default transport, so we don't really need to specify it here, and we never specify the transport when we override this value. Presuming that the transport is not present allows the value to be trivially parsed.
- The containers for the Pbench Agent and the Pbench Server are pushed from separate parts of our infrastructure -- the Agent is pushed via commands in agent/containers/images/Makefile while the Server is pushed directly from the Jenkins pipeline. This change adds an invocation of the new script after each of the push operations.
- The expiration is currently hard-coded at three weeks from when the container is pushed. We can consider parameterizing this if we care.
- Access to the repository is granted via a bearer token, which is stored in the Jenkins environment as another "credential" secret.