Refactor the authorization sub-system
The refactoring approach taken here covers two broad areas.  The first
being the removal of the `Auth` class replacing it with module
interfaces.  This change removes the need to pass the `Auth` instance
around between modules and objects, where a simple import of the `Auth`
module is sufficient.

The second area of refactoring is simplifying the authorization module
interfaces to hide knowledge of "internal tokens" vs OpenID Connect
tokens.  The refactoring solves a problem facing the code prior to this
change where one needs a working OpenID Connect instance to function,
but the fallback doesn't work to the internal token mechanism.

With these changes, if it is possible to talk to an OpenID Connect
instance, only OpenID Connect tokens will be accepted for authorization.
If that is not possible, only "internal" tokens will be accepted.  This
prevents the two code paths from having to co-exist seamlessly, such
that the unit test environment can explicitly operate in one or the
other.

For now, the unit test environment assumes internal tokens only.  Unit
tests for the OIDC code paths (with 100% coverage) are provided
separately.

Further, a number of other changes are worth noting explicitly:

 * The "authentication" section in the configuration file has been
   renamed to "openid-connect"

   There are four required options in that section required to use
   OpenID Connect for tokens: "server_url", "realm", "client", and
   "secret".  If the section or any of those options are missing then
   "internal" tokens will be expected.

   The `pbench-server-default.cfg` file now contains that section with
   the "server_url" option commented out since we are defaulting to
   "internal" tokens for the time being.

   The Pbench-in-a-Can `pbench-server.cfg` file was updated to use the
   OpenID Connect instance (KeyCloak) by defining the "server_url" and
   providing a "secret".

 * A new "flask-app" section was added to provide the "secret-key"
   option for populating the `Flask` app's `secret_key` property

 * The `Auth` module's `set_oidc_client` was moved to the OpenIDClient
   `create` class factory method to improve OIDC encapsulation

 * The `Auth` module's `get_user_id()` was renamed to
   `get_current_user_id()`

 * The `Auth` module's `verify_auth()` method calls one of two helper
   methods based on whether OpenID Connect or internal tokens are to
   be used

   The presence of the "openid-connect" section with all the options
   configured causes OpenID Connect to be used vs internal tokens

 * Knowledge of OIDC tokens pushed out of the `Auth` module and into the
   `OpenIDClient` class

   We did not make a separate module for internal tokens since the
   intention is to remove that code entire once we can depend on an
   OpenID Connect instance.

 * The `OpenIDClient` was refactored to further its encapsulation and
   to facilitate easier testing via its new `Connection` sister class

 * Finally, the unit tests for the `Auth` module proper were rewritten
   to leverage the new `Connection` class to make the test behaviors