Fix a few dashboard issues (#3261)
* Fix a few dashboard issues

PBENCH-993
PBENCH-1083

When I removed the `dataset.created` column as part of the `Sync` operational
changes, I neglected to correct dashboard code referencing `dataset.created`.
This should now be based on `dataset.uploaded` as that's the only date we are
guaranteed to have. (Especially once we start supporting archival of tarballs
without the Pbench Agent `metadata.log` format.) This corrects that oversight.

Since it's a quick fix, and I happened to run across the code in tracking down
all the `dataset.created` references, I also fixed the reversal of the
"Publish" and "Unpublish" action labels.