Filter datasets by metadata (#3213)
* Filter datasets by metadata

PBENCH-825

Create the new query parameter `?filter` on `/datasets/list` to allow the
query to filter on the basis of arbitrary `Metadata` key values. This code
experimentally allows a combination of _AND_ and _OR_ as well as exact string
and "contains" matches, using symbols in the filter value.

E.g., to find all "fio" runs from 2018, you could do the following. `:fio` is
an exact value match on "fio" while `:~2018` requires that the string "2018"
appear within the value:

 `?filter=dataset.metalog.pbench.script:fio,dataset.metalog.date:~2018`

(We have the `start` and `end` keywords, which provide more exact timestamp
matches, but you get the idea.)

Or, to find all datasets that have been favorited by the current user _or_
marked "seen" by the owner (should someone want to to this), the `^` character
provides _or_ linkage to consecutively designated values:

 `?filter=^user.dashboard.favorite:true,^global.dashboard.seen:false`

Or to find datasets that the authorized user has favorited, with a dataset
_name_ containing either "webb" or "dave":

 `?filter=user.dashboard.favorite:true,^dataset.name:~dave,^dataset.name:~webb`

This also introduces functional tests for filtered `/datasets/list`, which
rely on the `pytest-dependencies` package to ensure we run in the proper
sequence. (The one thing I _don't_ like here is that, because I need to push
`delete` after `list`, and eventually after other tests, it won't happen if
those tests fail. We may want to consider re-writing the delete test as a
`pytest` cleanup action so it happens at the end even if intermediate tests
fail... but I leave that as an exercise for some reader...)