Unified Pbench Server container deployment (#3260)
This change seeks to unify the respective processes of deployment of the
Pbench Server container for Staging and for functional testing:  with these
changes, the two deployments will now performed using the same script,
server/pbenchinacan/deploy.

The Staging deployment presupposes the existence of the services which
the Pbench Server depends upon, while in the functional test deployment
these services are provided by a pod which is started dynamically prior to
invoking the Pbench Server container. This pod previously was started by
run-pbench-in-a-can, but the code for this has been refactored into a new
script, deploy-dependencies which is now run as a subordinate script.
What remains for run-pbench-in-a-can, other than running the two
deployment scripts, is to set up the parameters required by
deploy-dependencies and deploy and to create the files and directory trees
which deploy will map into the Pbench Server container. (These files and
directory trees, for the most part, already exist in the Staging deployment.)

This change also introduces a number of artifacts used in the deployment:
rsyslog files, the Nginx favicon file, etc.

Finally, the details of the Pbench Server deployment have been redistributed
between the deploy script, the container-build.sh script, and the
run-pbench-in-a-can script. Configuring the files and directory trees for
the Server's use is now done by run-pbench-in-a-can instead of
container-build.sh; some details of container invocation are now configured
at build-time instead of via options on the podman run command, while the
remaining options have been harmonized between the two deployment
scenarios. And, the generation of the pbench-server.cfg file is now done
by run-pbench-in-a-can instead of container-build.sh.

So, container-build.sh now produces a single container image containing
only what is needed for a Staging deployment, and run-pbench-in-a-can
provides everything else needed to make the deploy script work for a
functional test deployment.