Add a Makefile for building and testing the Dashboard (#3273)
Add a Makefile for the Pbench Dashboard, to allow it to be built (and re-built) only as
needed and to provide an encapsulated interface for running lint and unit tests, etc.

The build.sh script is changed to use make to clear and install the Dashboard
dependences, to run lint, and to run unit tests. And, a new make invocation is
added after we build the Server and Agent RPMs to also build the Dashboard
deployment.

Also, jenkins/runlocal is modified to build the Dashboard, as well.