Rely on PUT API, remove indexing
Most of the files listed in this commit are removed.  Any part of the
code that touched Elasticsearch indexing has been removed, and all tests
related to indexing have been removed.

One very small change which has a big impact is the reliance on the
`pbench-results-push` Pbench Agent CLI command to "put" the tar ball
into the archive tree.

This code now relies on the code of the `main` branch running along side
it sharing the same `${ARCHIVE}` file system.