Merge `*-put-shim` into `*-dispatch`
Since we don't need a separate cronjob of the PUT API shim, merge that
functionality into `pbench-dispatch`.

The use of `pbench-results-push` only takes place when a `put-token`
option is configured in the `pbench-server` section of the configuration
file.  Each tar ball pushed to a Pbench Server via the PUT API always
includes the following metadata:

    {
      "global": {
        "server": {
          "version": "</opt/pbench-server/VERSION>",
          "sha1": "</opt/pbench-server/SHA1>",
          "hostname": "<$(hostname -f)>"
        }
      }
    }

When a tar ball is originally from a satellite server, the satellite
name is added as `server.origin` metadata.