Fold `*-backup-tarballs` into `*-dispatch`
We are not supporting S3 backups any time soon, and even if we have
plans to do so in the future, it won't happen in the b0.69 code base.

The "backup" story is really just a `cp -a` to an NFS mounted volume
for the most part.  As such, we strip out all the backup mechanism to
just a simple `cp -a` in `pbench-dispatch.sh` performed before we move
the tar ball to the normal ARCHIVE tree.

If the `cp -a` fails, we leave the tar ball where it is in the
reception area and do nothing else.

Remove `rsync` references where possible.