Correct alembic database state
The original `OperationName` ENUM type had the values BACKUP, DELETE, INDEX,
REINDEX, TOOLINDEX, UNPACK, and UPLOAD. The BACKUP and UNPACK operations were
later removed, and the UPDATE operation was added. However, alembic does not
check ENUM values nor does autogenerate provide automated upgrades.

Dropping ENUM values is problematic: it would mean generating code to query
the live server database and either drop or change existing rows, which would
be difficult to test. Having obsolete values in an ENUM is not problematic,
unless the number becomes overwhelming, so this PR doesn't attempt to address
that.

However, new server code will attempt to use the UPDATE operation, and this
will cause SQLAlchemy failures as the value is unknown. This PR simply adds
that value to the ENUM.