Passing server tests
The changes to the legacy server test gold files are numerous because
most tests require the use of `pbench-unpack-tarballs` to operate.

Many of the changes stem from the use of `logger` output, including the
new use of embedded JSON.

The mock'd up `logger` command is removed, along with the
`_testloggerpayload` references, and a `bash` `alias` is setup for
`logger` to cause it to emit to `stderr` instead of `syslog` for testing
purposes.

A majority of the gold file changes are for the `test-7.*.txt` files.
Since `pbench-index.py` is still issuing reports, the typical changes
are the removal of the `pbench-unpack-tarballs` report, the addition of
the `pbench-unpack-tarballs` syslog summary entry, and slight change in
the syslog entries for the `Report` class moved into the
`pbench-index.py` CLI interface.

Most of the remaining gold files are pure removals of the reports no
longer posted, and the removal of gratuitous log entries.