Remove `*-report-status` from `*-audit-server`
Because the `pbench-audit-server` command is used with every test, the
change to remove the use of `pbench-report-status` from that script
touches every legacy test gold file.

The jist of the change is two fold, 1) stop "reporting" status, 2)
maintain a report file locally in the `LOGSDIR` and just refer to it
in the logged data instead of including it in the log file itself.
There is no point repeating the audit report each time it runs.

During report generation, the previous report is moved aside in case
errors are encountered building the new one.