Drop crontab for indexers
PBENCH-1086

Run the indexer (and, for now, the re-indexer) as subprocesses controlled by
the Pbench Server rather than using crontab. The processes are monitored by a
`waitpid` thread that will report failure and restart, while each indexer runs
continuously waiting between polls for datasets that are READY.

Posting as a draft mostly because I need a break from unit testing. This has
been working on my VM, but with a patchwork container ... so I want to see how
it works "live". (Also, I'm archiving it because my new laptop is supposed to
arrive tomorrow and a "GitHub backup" seems prudent.)