Drop internal PUT size limit
PBENCH-151

We want to rely on easily configurable NGINX size limits for tarball uploads,
while the current server code checks an internal limit that defaults to 1Gb.

A quick survey of the production server shows over 5000 files exceeding 1G and
204 tarballs over 10G.

This PR simply removes the internal check. We never had a test for this since
we never found a way to fake the content-length header.