Support non-Pbench tarballs (#3317)
* Support non-Pbench tarballs

PBENCH-1077

We can envison taking advantage of the `server.archiveonly` option to utilize
the new Pbench Server to archive performance tarballs that weren't created by
Pbench Agent and lack the canonical directory structure and `metadata.log`;
but the current cache manager and `PUT` code fail if the file can't be found.

This PR allows successful upload when `metadata.log` isn't found, and adds a
functional test to verify. A tarball without `metadata.log` is implicitly
treated as `server.archiveonly:true` and therefore won't be indexed. As a
marker, we fabricate a `dataset.metalog` identifying it as a "Foreign" script.