Remove code for running on satellite instances
Our satellite support is maintained in `v0.69.10-server` (390f97925).
The `b0.69` branch can't support satellite operation because of the fact
that `pbench-dispatch` now unconditionally sends tar balls to a new
Pbench Server instance that runs along side it, and satellite's won't
have that.

The pieces of code that run only on satellites are:

 * `pbench-satellite-cleanup`
 * `pbench-satellite-state-change`
 * `pbench-sync-package-tarballs`

We use the `ssh` command mock in the legacy tests to mimic the behavior
of `pbench-sync-package-tarballs` and `pbench-satellite-state-change` as
it relates to the local operation of `pbench-sync-satellite`.

mock pbench-sync-package-tarballs