Add checks for all required sync satellite configs
Before beginning any `pbench-sync-satellite` operation, each config
option is checked before beginning.  The normal logging is setup first
so that it will also show up in the syslog entries.

The new `test-0.3` covers the verification of the required backup
directory configuration for `pbench-dispatch`, and the new `test-0.4`
covers the verification of the required `pbench-receive-dir-prefix` for
pbench-sync-satellite & pbench-dispatch.

The new `test-9` covers a missing argument to `pbench-sync-satellite`,
and the tests `test-29` .. `test-32` covers handling of the four
required satellite options.