Drop `ARCHIVE` for simple "unpack" directory
The goal of this work is to have the `b0.69` branch handle backups of
all tar balls, archiving via the PUT API (`pbench-results-push`), and
unpacking the tar ball into the INCOMING, RESULTS, and USERS trees.

The ARCHIVE tree is no longer required since there is no need to access
the tar ball once it is backed up and unpacked.  The old method of using
state directories is replaced with a simple "unpack" directory where the
tar ball is placed in that directory when it is ready to be unpacked,
and removed once it is successfully unpacked.

The removal of the ARCHIVE tree, however, has a cascading effect as it
touches many files, mostly the "gold" files of the tests.

Some particular notes on various cron jobs and interfaces:

 * `pbench-dispatch`
   * Re-worked to use an unpack directory tree instead of symbolic links
     to track operational state
 * `pbench-unpack-tarballs`
   * Reworked to pull tar balls from the unpack directories
 * `pbench-audit-server`
   * This is mostly a removal of the entire report on the ARCHIVE tree,
     but also removed validation against the ARCHIVE
 * `pbench-check-tb-age`
   * References were for validation, dropped
 * `pbench-clean-up-dangling-results-links`
   * Dropped entirely, as this code was only partially implemented
 * `pbench-cull-unpacked-tarballs`
   * References were for validation, dropped
 * `pbench-restore-unpacked-tarballs`
   * Switched to just using the backup tree
 * `pbench-sync-satellite`
   * No impact, did not require the ARCHIVE tree since it feeds dispatch
 * `pbench-tarball-stats`
   * Changed to use the backup tree instead