More tweakies ...
So, forget I even thought about dealing with the `MetadataError` here. I can't
collect "errors" here because `_get_dataset_metadata` actually internally
traps any "real" metadata error and returns a `None` value already. To collect
a list of keyspace errors I'd need to change `_get_dataset_metadata`, and I'm
not going to do that here.

The only `MetadataError` that can get through here is actually a bad keyname,
and we've already done that validation so we really don't need to worry about
it.

I'd like to revise this separately because I've realized that while I allow
`filter` to reference the weird non-alphanumeric key paths in `metalog`, I
didn't do that for the `metadata` parameter... and that'll affect the
`_get_dataset_metadata` as well because it independently validates the key
paths and wouldn't know about that exception. (Ugh.) Not here. I did however
add a test case showing that a keyspace conflict error results in a `None`
value.