Move set-expiration of Server container to a conditional pipeline stage
The container images built by the CI are pushed to an external registry and
set with an expiration date for auto-deletion.  This is valuable for
transient images which are built for PRs, but it is unfortunate for images
built from branches in the main repo, such as release branches and `main`.
This change moves the setting of the expiration date to its own pipeline
stage which is executed only for PRs, so that branch builds will not expire.

PBENCH-1113