Fix /datasets?filter to select across namespaces
PBENCH-1117

I discovered that the single simplistic `LEFT JOIN` allows combining native
`Dataset` and `Metadata` terms in a `SELECT`, but with limitations: because
the SQL join constructs a row for each `Metadata` match, matches for, e.g.,
`server.origin` and `dataset.metalog.pbench.script` or `global.server.legacy`
will appear on separate table rows. Each has duplicate `Dataset` columns, but
that doesn't help when trying to select across namespaces.

The only effective solution I was able to find was to cascade the joins in
order to build a new table with a separate column for each metadata namespace
row matching the dataset. This allows a single `SELECT` to work across the
columns in the new table.

I'm somewhat unhappy with the `filter` unit tests, because the `sqlite3`
engine seems to be generating odd and incorrect SQL for queries involving the
`user` namespace. (PostgreSQL does this correctly on a live server.) This
should be a fourth `LEFT JOIN` but on a user_id match as well as the dataset
reference and primary key name. I've made the unit tests "pass" ... but the
SQL being compared wouldn't actually work correctly. (Also note that the uery
builder omits the `user` join if the query doesn't reference that alias, but
doesn't similarly optimize for the other joins, which seems weird.)