Fix /datasets?filter to select across namespaces (#3359)
* Fix /datasets?filter to select across namespaces

PBENCH-1117

I discovered that the single simplistic `LEFT JOIN` allows combining native
`Dataset` and `Metadata` terms in a `SELECT`, but with limitations: because
the SQL join constructs a row for each `Metadata` match, matches for, e.g.,
`server.origin` and `dataset.metalog.pbench.script` or `global.server.legacy`
will appear on separate table rows. Each has duplicate `Dataset` columns, but
that doesn't help when trying to select across namespaces.

The only effective solution I was able to find was to cascade the joins in
order to build a new table with a separate column for each metadata namespace
row matching the dataset. This allows a single `SELECT` to work across the
columns in the new table.