Remove simplistic "API" list from endpoints
PBENCH-1119

With the REST-ification of the API, only a few of the APIs can still be called
using the simplistic `{name: uri}` list, because URI parameters are needed and
are no longer at the end of the URI: for example,
`/api/v1/datasets/{dataset}/metadata`.

The UI now has the `uriTemplate` method to help expand the `"uri"` template
objects. Complete the transition by updating the remaining UI references and
dropping the `api` object entirely.